[YSJ] - 동아리 정보에 맞게 이미지 슬라이더 생성되도록 수정 - 변수명 / 경로명 수정
